# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Integrador API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Lista todos los pacientes",
                "description": "Obtiene una lista de todos los pacientes disponibles.",
                "operationId": "5d245d8dbe3006df98a6327d0ef92f80",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Patient"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Crea un nuevo paciente",
                "description": "Crea un nuevo paciente con la información proporcionada.",
                "operationId": "990c6f96dd7d4d2b3b6b2159d31b3cbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/patients/{id}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtiene un paciente por ID",
                "description": "Obtiene los detalles de un paciente específico por su ID.",
                "operationId": "7ddcd48af8cd38331a0526024049c344",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Actualiza un paciente por ID",
                "description": "Actualiza la información de un paciente específico por su ID.",
                "operationId": "f60dd714fa74bbfa62f73399a3ff4026",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Elimina un paciente por ID",
                "description": "Elimina un paciente específico por su ID.",
                "operationId": "f63ab009b8eb47990a2399733b06b070",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operación exitosa (sin contenido)"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePatientRequest": {
                "title": "Store Patient Request",
                "description": "Store Patient request body data",
                "required": [
                    "nombre",
                    "fecha_nacimiento",
                    "direccion",
                    "dni",
                    "sip",
                    "telefono",
                    "email",
                    "zona_id"
                ],
                "properties": {
                    "nombre": {
                        "description": "Patient name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "fecha_nacimiento": {
                        "description": "Patient birth date",
                        "type": "string",
                        "format": "date"
                    },
                    "direccion": {
                        "description": "Patient address",
                        "type": "string",
                        "maxLength": 255
                    },
                    "dni": {
                        "description": "Patient DNI",
                        "type": "string",
                        "maxLength": 20
                    },
                    "sip": {
                        "description": "Patient SIP",
                        "type": "integer"
                    },
                    "telefono": {
                        "description": "Patient phone number",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Patient email",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "zona_id": {
                        "description": "Zone ID",
                        "type": "integer"
                    },
                    "contacto_id": {
                        "description": "Contact ID",
                        "type": "integer"
                    },
                    "situacion_personal": {
                        "description": "Patient personal situation",
                        "type": "string"
                    },
                    "situacion_sanitaria": {
                        "description": "Patient health situation",
                        "type": "string"
                    },
                    "situacion_habitage": {
                        "description": "Patient housing situation",
                        "type": "string"
                    },
                    "autonomia": {
                        "description": "Patient autonomy level",
                        "type": "string"
                    },
                    "situacion_economica": {
                        "description": "Patient economic situation",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdatePatientRequest": {
                "title": "Update Patient Request",
                "description": "Update Patient request body data",
                "properties": {
                    "nombre": {
                        "description": "Patient name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "fecha_nacimiento": {
                        "description": "Patient birth date",
                        "type": "string",
                        "format": "date"
                    },
                    "direccion": {
                        "description": "Patient address",
                        "type": "string",
                        "maxLength": 255
                    },
                    "dni": {
                        "description": "Patient DNI",
                        "type": "string",
                        "maxLength": 20
                    },
                    "sip": {
                        "description": "Patient SIP",
                        "type": "integer"
                    },
                    "telefono": {
                        "description": "Patient phone number",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Patient email",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "zona_id": {
                        "description": "Zone ID",
                        "type": "integer"
                    },
                    "contacto_id": {
                        "description": "Contact ID",
                        "type": "integer"
                    },
                    "situacion_personal": {
                        "description": "Patient personal situation",
                        "type": "string"
                    },
                    "situacion_sanitaria": {
                        "description": "Patient health situation",
                        "type": "string"
                    },
                    "situacion_habitage": {
                        "description": "Patient housing situation",
                        "type": "string"
                    },
                    "autonomia": {
                        "description": "Patient autonomy level",
                        "type": "string"
                    },
                    "situacion_economica": {
                        "description": "Patient economic situation",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Patient": {
                "title": "Patient",
                "description": "Patient model",
                "properties": {
                    "id": {
                        "description": "Patient ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Patient name",
                        "type": "string"
                    },
                    "fecha_nacimiento": {
                        "description": "Patient birth date",
                        "type": "string",
                        "format": "date"
                    },
                    "direccion": {
                        "description": "Patient address",
                        "type": "string"
                    },
                    "dni": {
                        "description": "Patient DNI",
                        "type": "string"
                    },
                    "sip": {
                        "description": "Patient SIP",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Patient phone number",
                        "type": "string"
                    },
                    "email": {
                        "description": "Patient email",
                        "type": "string",
                        "format": "email"
                    },
                    "zona_id": {
                        "description": "Zone ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "contacto_id": {
                        "description": "Contact ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "situacion_personal": {
                        "description": "Patient personal situation",
                        "type": "string"
                    },
                    "situacion_sanitaria": {
                        "description": "Patient health situation",
                        "type": "string"
                    },
                    "situacion_habitage": {
                        "description": "Patient housing situation",
                        "type": "string"
                    },
                    "autonomia": {
                        "description": "Patient autonomy level",
                        "type": "string"
                    },
                    "situacion_economica": {
                        "description": "Patient economic situation",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}